+++ playground/src/tabs/Aria.tsx
@@ -1,15 +1,18 @@
 import React from "react";
-import { View, Text, Flex, ChipGroup, Item } from "@atomicjolt/atomic-elements";
+import { View, Text, Flex } from "@atomicjolt/atomic-elements";
 
 export default function Aria() {
   return (
-    <ChipGroup label="Label">
-      <Item>
-        <Text>Chip 1</Text>
-      </Item>
-      <Item>
-        <Text>Chip 2</Text>
-      </Item>
-    </ChipGroup>
+    <Flex $direction="column" $gap="3">
+      <Text $size="1">The quick brown fox jumps over the lazy dog.</Text>
+      <Text $size="2">The quick brown fox jumps over the lazy dog.</Text>
+      <Text $size="3">The quick brown fox jumps over the lazy dog.</Text>
+      <Text $size="4">The quick brown fox jumps over the lazy dog.</Text>
+      <Text $size="5">The quick brown fox jumps over the lazy dog.</Text>
+      <Text $size="6">The quick brown fox jumps over the lazy dog.</Text>
+      <Text $size="7">The quick brown fox jumps over the lazy dog.</Text>
+      <Text $size="8">The quick brown fox jumps over the lazy dog.</Text>
+      <Text $size="9">The quick brown fox jumps over the lazy dog.</Text>
+    </Flex>
   );
 }